asr/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Audio processing and speech recognition
numpy==1.24.3
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
librosa==0.10.1
soundfile==0.12.1

# File handling
python-multipart==0.0.6
aiofiles==23.2.0

# Monitoring and metrics
prometheus-client==0.19.0

# Configuration
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0